feat: move preset entities
