--- multiauth/new/entities/__init__.py
@@ -1,1 +1,0 @@
-"""Types and entities for multiauth."""
