+++ multiauth/revamp/lib/presets/aws.py
@@ -2,7 +2,7 @@ import enum
 
 from pydantic import Field
 
-from multiauth.new.entities.main import AuthProvider, AuthRequester, Credentials
+from multiauth.revamp.lib.presets.main import AuthProvider, AuthRequester, Credentials
 
 ###########################
 ## TODO(maxence@escape, antoine@escape): We should stop doing that and use this method: https://stackoverflow.com/questions/58833462/aws-cognito-authentication-curl-call-generate-token-without-cli-no-clien
